feat: add plaintext tag
allows outputting content only in the plaintext version
